test(ddd-kitchenpos): split fixture classes
